+++ policyengine-client/src/App.jsx
@@ -11,11 +11,13 @@ import PolicyEngine from "./policyengine/policyengine";
 import createRedirects from "./policyengine/tools/namedPolicies";
 import { UK, US } from "./countries";
 import MarkdownPage from "./policyengine/pages/markdown";
+import LandingPage from "./landing";
 
 // Markdown files
 
 import UK_FAQ from "./markdown/uk/faq.md";
 import ABOUT from "./markdown/about.md";
+import CONTACT from "./markdown/contact.md";
 // Import other markdown files here
 
 
@@ -25,6 +27,7 @@ export default function App(props) {
     const markdownPages = [
         { content: UK_FAQ, path: "/uk/faq", title: "FAQ" },
         { content: ABOUT, path: "/about", title: "About" },
+        { content: CONTACT, path: "/contact", title: "Contact" },
         // Add other pages here
     ];
     return (
@@ -36,7 +39,7 @@ export default function App(props) {
                     </Route>
                 ))}
                 <Route exact path="/">
-                    <Redirect to="/uk/policy" />
+                    <LandingPage />
                 </Route>
                 <Route exact path="/uk">
                     <Redirect to="/uk/policy" />