+++ policyengine-client/src/policyengine/pages/policy/policy.jsx
@@ -2,9 +2,11 @@ import React, { useContext } from "react"
 import { Col, Row } from "react-bootstrap";
 import { CountryContext } from "../../../countries"
 import Menu from "./menu";
-import { OverviewHolder, PolicyOverview } from "./overview";
+import { OverviewHolder, PolicyOverview, SharePolicyLinks } from "./overview";
 import Parameter from "./parameter";
 import NavigationButton from "../../general/navigationButton";
+import { Affix } from "antd";
+import HelpButton from "../../general/help";
 
 
 export default class Policy extends React.Component {
@@ -32,6 +34,7 @@ export default class Policy extends React.Component {
 
     render() {
         return <>
+            <HelpButton />
             <Row>
                 <Col xl={3}>
                     <Menu
@@ -46,23 +49,28 @@ export default class Policy extends React.Component {
                 </Col>
                 <Col xl={3}>
                     <OverviewHolder>
-                    <PolicyOverview page="policy"/>
-                    <div className="d-block align-middle">
-                        <div className="d-flex justify-content-center">
-                            {this.context.showPopulationImpact && <NavigationButton 
-                                primary 
-                                target="population-impact" 
-                                text={`Calculate ${this.context.properName} impact`} 
-                            />}
-                        </div>
-                        <div className="d-flex justify-content-center">
-                            {this.context.showHousehold && <NavigationButton
-                                target="household" 
-                                text="Describe your household"
-                                primary={!this.context.showPopulationImpact}
-                            />}
-                        </div>
-                    </div>
+                        <Affix offsetTop={55}>
+                            <PolicyOverview />
+				        </Affix>
+                        <Affix offsetTop={400}>
+                            <SharePolicyLinks page="policy"/>
+                            <div className="d-block align-middle">
+                                <div className="d-flex justify-content-center">
+                                    {this.context.showPopulationImpact && <NavigationButton 
+                                        primary 
+                                        target="population-impact" 
+                                        text={`Calculate ${this.context.properName} impact`} 
+                                    />}
+                                </div>
+                                <div className="d-flex justify-content-center">
+                                    {this.context.showHousehold && <NavigationButton
+                                        target="household" 
+                                        text="Describe your household"
+                                        primary={!this.context.showPopulationImpact}
+                                    />}
+                                </div>
+                            </div>
+                        </Affix>
                     </OverviewHolder>
                 </Col>
             </Row>